Release/3.3.2

See merge request arkcase/ArkCase!1679

+++ acm-services/acm-service-sequence-manager/src/main/java/com/armedia/acm/services/sequence/listener/AcmSequenceDatabaseChangeHandler.java
@@ -30,12 +30,14 @@ package com.armedia.acm.services.sequence.listener;
 import com.armedia.acm.data.AcmDatabaseChangesEvent;
 import com.armedia.acm.data.AcmEntity;
 import com.armedia.acm.data.AcmObjectChangelist;
+import com.armedia.acm.services.sequence.annotation.AcmSequence;
 import com.armedia.acm.services.sequence.annotation.AcmSequenceAnnotationReader;
+import com.armedia.acm.services.sequence.generator.AcmSequenceGeneratorManager;
 import com.armedia.acm.services.sequence.service.AcmSequenceService;
 
 import org.apache.commons.beanutils.PropertyUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.apache.logging.log4j.Logger;
+import org.apache.logging.log4j.LogManager;
 import org.springframework.context.ApplicationListener;
 
 import java.lang.reflect.Field;
@@ -47,9 +49,11 @@ import java.util.List;
  */
 public class AcmSequenceDatabaseChangeHandler implements ApplicationListener<AcmDatabaseChangesEvent>
 {
-    private transient Logger log = LoggerFactory.getLogger(getClass());
+    private transient Logger log = LogManager.getLogger(getClass());
 
     private AcmSequenceAnnotationReader sequenceAnnotationReader;
+
+    private AcmSequenceGeneratorManager sequenceGeneratorManager;
 
     private AcmSequenceService sequenceService;
 
@@ -72,8 +76,12 @@ public class AcmSequenceDatabaseChangeHandler implements ApplicationListener<Acm
                 {
                     try
                     {
-                        String sequenceValue = PropertyUtils.getProperty(object, annotatedField.getName()).toString();
-                        getSequenceService().removeSequenceRegistry(sequenceValue);
+                        String sequenceName = annotatedField.getAnnotation(AcmSequence.class).sequenceName();
+                        if (getSequenceGeneratorManager().isSequenceEnabled(sequenceName))
+                        {
+                            String sequenceValue = PropertyUtils.getProperty(object, annotatedField.getName()).toString();
+                            getSequenceService().removeSequenceRegistry(sequenceValue);
+                        }
                     }
                     catch (Exception e)
                     {
@@ -100,6 +108,23 @@ public class AcmSequenceDatabaseChangeHandler implements ApplicationListener<Acm
     public void setSequenceAnnotationReader(AcmSequenceAnnotationReader sequenceAnnotationReader)
     {
         this.sequenceAnnotationReader = sequenceAnnotationReader;
+    }
+
+    /**
+     * @return the sequenceGeneratorManager
+     */
+    public AcmSequenceGeneratorManager getSequenceGeneratorManager()
+    {
+        return sequenceGeneratorManager;
+    }
+
+    /**
+     * @param sequenceGeneratorManager
+     *            the sequenceGeneratorManager to set
+     */
+    public void setSequenceGeneratorManager(AcmSequenceGeneratorManager sequenceGeneratorManager)
+    {
+        this.sequenceGeneratorManager = sequenceGeneratorManager;
     }
 
     /**
